Control.Monad.Primitive

Description

Primitive state-transformer monads.

Documentation

class Monad m => PrimMonad m where

Class of monads which can perform primitive state-transformer actions.

Associated Types

type PrimState m

State token type.

Methods

primitive :: (State# (PrimState m) -> (# State# (PrimState m), a #)) -> m a

Execute a primitive operation.

data RealWorld

RealWorld is deeply magical. It is primitive, but it is not unlifted (hence ptrArg). We never manipulate values of type RealWorld; it's only used in the type system, to parameterise State#.

primitive_ :: PrimMonad m => (State# (PrimState m) -> State# (PrimState m)) -> m ()

Execute a primitive operation with no result.

class PrimMonad m => PrimBase m where

Class of primitive monads for state-transformer actions.

Unlike PrimMonad, this typeclass requires that the Monad be fully expressed as a state transformer, therefore disallowing other monad transformers on top of the base IO or ST.

Since: 0.6.0.0

Methods

internal :: m a -> State# (PrimState m) -> (# State# (PrimState m), a #)

Expose the internal structure of the monad.

class (PrimMonad m, s ~ PrimState m) => MonadPrim s m

PrimMonad's state token type can be annoying to handle in constraints. This typeclass lets users (visually) notice PrimState equality constraints less, by witnessing that s ~ PrimState m.

class (PrimBase m, MonadPrim s m) => MonadPrimBase s m

PrimBase's state token type can be annoying to handle in constraints. This typeclass lets users (visually) notice PrimState equality constraints less, by witnessing that s ~ PrimState m.

liftPrim :: (PrimBase m1, PrimMonad m2, PrimState m1 ~ PrimState m2) => m1 a -> m2 a

Lifts a PrimBase into another PrimMonad with the same underlying state token type.

primToPrim :: (PrimBase m1, PrimMonad m2, PrimState m1 ~ PrimState m2) => m1 a -> m2 a

Convert a PrimBase to another monad with the same state token.

primToIO :: (PrimBase m, PrimState m ~ RealWorld) => m a -> IO a

Convert a PrimBase with a RealWorld state token to IO

primToST :: PrimBase m => m a -> ST (PrimState m) a

Convert a PrimBase to ST

ioToPrim :: (PrimMonad m, PrimState m ~ RealWorld) => IO a -> m a

Convert an IO action to a PrimMonad.

Since: 0.6.2.0

stToPrim :: PrimMonad m => ST (PrimState m) a -> m a

Convert an ST action to a PrimMonad.

Since: 0.6.2.0

unsafePrimToPrim :: (PrimBase m1, PrimMonad m2) => m1 a -> m2 a

Convert a PrimBase to another monad with a possibly different state token. This operation is highly unsafe!

unsafePrimToIO :: PrimBase m => m a -> IO a

Convert any PrimBase to IO. This operation is highly unsafe!

unsafePrimToST :: PrimBase m => m a -> ST s a

Convert any PrimBase to ST with an arbitrary state token. This operation is highly unsafe!

unsafeIOToPrim :: PrimMonad m => IO a -> m a

Convert an IO action to any PrimMonad. This operation is highly unsafe!

Since: 0.6.2.0

unsafeSTToPrim :: PrimMonad m => ST s a -> m a

Convert an ST action with an arbitrary state token to any PrimMonad. This operation is highly unsafe!

Since: 0.6.2.0

unsafeInlinePrim :: PrimBase m => m a -> a

See unsafeInlineIO. This function is not recommended for the same reasons.

unsafeInlineIO :: IO a -> a

Generally, do not use this function. It is the same as accursedUnutterablePerformIO from bytestring and is well behaved under narrow conditions. See the documentation of that function to get an idea of when this is sound. In most cases GHC.IO.Unsafe.unsafeDupablePerformIO should be preferred.

unsafeInlineST :: ST s a -> a

See unsafeInlineIO. This function is not recommended for the same reasons. Prefer runST when s is free.

touch :: PrimMonad m => a -> m ()

keepAlive :: PrimBase m => a -> (a -> m r) -> m r

evalPrim :: forall a m. PrimMonad m => a -> m a

Create an action to force a value; generalizes evaluate

Since: 0.6.2.0

unsafeInterleave :: PrimBase m => m a -> m a

unsafeDupableInterleave :: PrimBase m => m a -> m a

noDuplicate :: PrimMonad m => m ()