Data.Primitive.Ptr

Description

Primitive operations on machine addresses.

Since: 0.6.4.0

Types

data Ptr a

Constructors

Ptr Addr#

Address arithmetic

nullPtr :: Ptr a

advancePtr :: Prim a => Ptr a -> Int -> Ptr a

Offset a pointer by the given number of elements.

subtractPtr :: Prim a => Ptr a -> Ptr a -> Int

Subtract a pointer from another pointer. The result represents the number of elements of type a that fit in the contiguous memory range bounded by these two pointers.

Element access

indexOffPtr :: Prim a => Ptr a -> Int -> a

Read a value from a memory position given by a pointer and an offset. The memory block the address refers to must be immutable. The offset is in elements of type a rather than in bytes.

readOffPtr :: (Prim a, PrimMonad m) => Ptr a -> Int -> m a

Read a value from a memory position given by an address and an offset. The offset is in elements of type a rather than in bytes.

writeOffPtr :: (Prim a, PrimMonad m) => Ptr a -> Int -> a -> m ()

Write a value to a memory position given by an address and an offset. The offset is in elements of type a rather than in bytes.

Block operations

copyPtr

Arguments

:: (PrimMonad m, Prim a)

=> Ptr a
destination pointer

-> Ptr a
source pointer

-> Int
number of elements

-> m ()

Copy the given number of elements from the second Ptr to the first. The areas may not overlap.

movePtr

Arguments

:: (PrimMonad m, Prim a)

=> Ptr a
destination pointer

-> Ptr a
source pointer

-> Int
number of elements

-> m ()

Copy the given number of elements from the second Ptr to the first. The areas may overlap.

setPtr :: (Prim a, PrimMonad m) => Ptr a -> Int -> a -> m ()

Fill a memory block with the given value. The length is in elements of type a rather than in bytes.

copyPtrToMutablePrimArray

Arguments

:: (PrimMonad m, Prim a)

=> MutablePrimArray (PrimState m) a
destination array

-> Int
destination offset

-> Ptr a
source pointer

-> Int
number of elements

-> m ()

Copy from a pointer to a mutable primitive array. The offset and length are given in elements of type a. This function assumes that the Prim instance of a agrees with the Storable instance.

Note: this function does not do bounds or overlap checking.

copyPtrToMutableByteArray

Arguments

:: (PrimMonad m, Prim a)

=> MutableByteArray (PrimState m)
destination array

-> Int
destination offset given in elements of type a

-> Ptr a
source pointer

-> Int
number of elements

-> m ()

Copy from an unmanaged pointer address to a byte array. These must not overlap. The offset and length are given in elements, not in bytes.

Note: this function does not do bounds or overlap checking.