Data.Primitive.SmallArray

Description

Small arrays are boxed (im)mutable arrays.

The underlying structure of the Array type contains a card table, allowing segments of the array to be marked as having been mutated. This allows the garbage collector to only re-traverse segments of the array that have been marked during certain phases, rather than having to traverse the entire array.

SmallArray lacks this table. This means that it takes up less memory and has slightly faster writes. It is also more efficient during garbage collection so long as the card table would have a single entry covering the entire array. These advantages make them suitable for use as arrays that are known to be small.

The card size is 128, so for uses much larger than that, Array would likely be superior.

Documentation

data SmallArray a

Constructors

SmallArray (SmallArray# a)

data SmallMutableArray s a

Constructors

SmallMutableArray (SmallMutableArray# s a)

newSmallArray

Arguments

:: PrimMonad m

=> Int
size

-> a
initial contents

-> m (SmallMutableArray (PrimState m) a)

Create a new small mutable array.

Note: this function does not check if the input is non-negative.

readSmallArray

Arguments

:: PrimMonad m

=> SmallMutableArray (PrimState m) a
array

-> Int
index

-> m a

Read the element at a given index in a mutable array.

Note: this function does not do bounds checking.

writeSmallArray

Arguments

:: PrimMonad m

=> SmallMutableArray (PrimState m) a
array

-> Int
index

-> a
new element

-> m ()

Write an element at the given idex in a mutable array.

Note: this function does not do bounds checking.

copySmallArray

Arguments

:: PrimMonad m

=> SmallMutableArray (PrimState m) a
destination

-> Int
destination offset

-> SmallArray a
source

-> Int
source offset

-> Int
length

-> m ()

Copy a slice of an immutable array into a mutable array.

Note: this function does not do bounds or overlap checking.

copySmallMutableArray

Arguments

:: PrimMonad m

=> SmallMutableArray (PrimState m) a
destination

-> Int
destination offset

-> SmallMutableArray (PrimState m) a
source

-> Int
source offset

-> Int
length

-> m ()

Copy a slice of one mutable array into another.

Note: this function does not do bounds or overlap checking.

indexSmallArray

Arguments

:: SmallArray a
array

-> Int
index

-> a

Look up an element in an immutable array.

Note: this function does not do bounds checking.

indexSmallArrayM

Arguments

:: Monad m

=> SmallArray a
array

-> Int
index

-> m a

Look up an element in an immutable array.

The purpose of returning a result using a monad is to allow the caller to avoid retaining references to the array. Evaluating the return value will cause the array lookup to be performed, even though it may not require the element of the array to be evaluated (which could throw an exception). For instance:

data Box a = Box a
...

f sa = case indexSmallArrayM sa 0 of
  Box x -> ...

x is not a closure that references sa as it would be if we instead wrote:

let x = indexSmallArray sa 0

It also does not prevent sa from being garbage collected.

Note that Identity is not adequate for this use, as it is a newtype, and cannot be evaluated without evaluating the element.

Note: this function does not do bounds checking.

indexSmallArray## :: SmallArray a -> Int -> (# a #)

Read a value from the immutable array at the given index, returning the result in an unboxed unary tuple. This is currently used to implement folds.

Note: this function does not do bounds checking.

cloneSmallArray

Arguments

:: SmallArray a
source

-> Int
offset

-> Int
length

-> SmallArray a

Create a copy of a slice of an immutable array.

Note: The provided array should contain the full subrange specified by the two Ints, but this is not checked.

cloneSmallMutableArray

Arguments

:: PrimMonad m

=> SmallMutableArray (PrimState m) a
source

-> Int
offset

-> Int
length

-> m (SmallMutableArray (PrimState m) a)

Create a copy of a slice of a mutable array.

Note: The provided array should contain the full subrange specified by the two Ints, but this is not checked.

freezeSmallArray

Arguments

:: PrimMonad m

=> SmallMutableArray (PrimState m) a
source

-> Int
offset

-> Int
length

-> m (SmallArray a)

Create an immutable array corresponding to a slice of a mutable array.

This operation copies the portion of the array to be frozen.

Note: The provided array should contain the full subrange specified by the two Ints, but this is not checked.

unsafeFreezeSmallArray :: PrimMonad m => SmallMutableArray (PrimState m) a -> m (SmallArray a)

Render a mutable array immutable.

This operation performs no copying, so care must be taken not to modify the input array after freezing.

thawSmallArray

Arguments

:: PrimMonad m

=> SmallArray a
source

-> Int
offset

-> Int
length

-> m (SmallMutableArray (PrimState m) a)

Create a mutable array corresponding to a slice of an immutable array.

This operation copies the portion of the array to be thawed.

Note: The provided array should contain the full subrange specified by the two Ints, but this is not checked.

unsafeThawSmallArray :: PrimMonad m => SmallArray a -> m (SmallMutableArray (PrimState m) a)

Render an immutable array mutable.

This operation performs no copying, so care must be taken with its use.

runSmallArray :: (forall s. ST s (SmallMutableArray s a)) -> SmallArray a

Execute the monadic action and freeze the resulting array.

runSmallArray m = runST $ m >>= unsafeFreezeSmallArray

createSmallArray :: Int -> a -> (forall s. SmallMutableArray s a -> ST s ()) -> SmallArray a

Create an array of the given size with a default value, apply the monadic function and freeze the result. If the size is 0, return emptySmallArray (rather than a new copy thereof).

createSmallArray 0 _ _ = emptySmallArray
createSmallArray n x f = runSmallArray $ do
  mary <- newSmallArray n x
  f mary
  pure mary

sizeofSmallArray :: SmallArray a -> Int

The number of elements in an immutable array.

sizeofSmallMutableArray :: SmallMutableArray s a -> Int

The number of elements in a mutable array.

shrinkSmallMutableArray :: PrimMonad m => SmallMutableArray (PrimState m) a -> Int -> m ()

Shrink the mutable array in place. The size given must be equal to or less than the current size of the array. This is not checked.

resizeSmallMutableArray

Arguments

:: PrimMonad m

=> SmallMutableArray (PrimState m) a

-> Int
New size

-> a
Newly created slots initialized to this element. Only used when array is grown.

-> m (SmallMutableArray (PrimState m) a)

Resize a mutable array to new specified size. The returned SmallMutableArray is either the original SmallMutableArray resized in-place or, if not possible, a newly allocated SmallMutableArray with the original content copied over.

To avoid undefined behaviour, the original SmallMutableArray shall not be accessed anymore after a resizeSmallMutableArray has been performed. Moreover, no reference to the old one should be kept in order to allow garbage collection of the original SmallMutableArray in case a new SmallMutableArray had to be allocated.

emptySmallArray :: SmallArray a

The empty SmallArray.

smallArrayFromList :: [a] -> SmallArray a

Create a SmallArray from a list.

smallArrayFromListN :: Int -> [a] -> SmallArray a

Create a SmallArray from a list of a known length. If the length of the list does not match the given length, this throws an exception.

mapSmallArray' :: (a -> b) -> SmallArray a -> SmallArray b

Strict map over the elements of the array.

traverseSmallArrayP :: PrimMonad m => (a -> m b) -> SmallArray a -> m (SmallArray b)

This is the fastest, most straightforward way to traverse an array, but it only works correctly with a sufficiently "affine" PrimMonad instance. In particular, it must only produce one result array. ListT-transformed monads, for example, will not work right at all.