Data.Primitive.Types

Description

Basic types and classes for primitive array operations.

Documentation

class Prim a where

Class of types supporting primitive array operations. This includes interfacing with GC-managed memory (functions suffixed with ByteArray#) and interfacing with unmanaged memory (functions suffixed with Addr#). Endianness is platform-dependent.

Minimal complete definition

(sizeOfType# | sizeOf#), (alignmentOfType# | alignment#), indexByteArray#, readByteArray#, writeByteArray#, indexOffAddr#, readOffAddr#, writeOffAddr#

Methods

sizeOfType# :: Proxy a -> Int#

The size of values of type a in bytes. This has to be used with TypeApplications: sizeOfType @a.

Since: 0.9.0.0

sizeOf# :: a -> Int#

The size of values of type a in bytes. The argument is not used.

It is recommended to use sizeOfType# instead.

alignmentOfType# :: Proxy a -> Int#

The alignment of values of type a in bytes. This has to be used with TypeApplications: alignmentOfType @a.

Since: 0.9.0.0

alignment# :: a -> Int#

The alignment of values of type a in bytes. The argument is not used.

It is recommended to use alignmentOfType# instead.

indexByteArray# :: ByteArray# -> Int# -> a

Read a value from the array. The offset is in elements of type a rather than in bytes.

readByteArray# :: MutableByteArray# s -> Int# -> State# s -> (# State# s, a #)

Read a value from the mutable array. The offset is in elements of type a rather than in bytes.

writeByteArray# :: MutableByteArray# s -> Int# -> a -> State# s -> State# s

Write a value to the mutable array. The offset is in elements of type a rather than in bytes.

setByteArray#

Arguments

:: MutableByteArray# s

-> Int#
offset

-> Int#
length

-> a

-> State# s

-> State# s

Fill a slice of the mutable array with a value. The offset and length of the chunk are in elements of type a rather than in bytes.

indexOffAddr# :: Addr# -> Int# -> a

Read a value from a memory position given by an address and an offset. The memory block the address refers to must be immutable. The offset is in elements of type a rather than in bytes.

readOffAddr# :: Addr# -> Int# -> State# s -> (# State# s, a #)

Read a value from a memory position given by an address and an offset. The offset is in elements of type a rather than in bytes.

writeOffAddr# :: Addr# -> Int# -> a -> State# s -> State# s

Write a value to a memory position given by an address and an offset. The offset is in elements of type a rather than in bytes.

setOffAddr#

Arguments

:: Addr#

-> Int#
offset

-> Int#
length

-> a

-> State# s

-> State# s

Fill a memory block given by an address, an offset and a length. The offset and length are in elements of type a rather than in bytes.

sizeOf :: Prim a => a -> Int

The size of values of type a in bytes. The argument is not used.

It is recommended to use sizeOfType instead.

This function has existed since 0.1, but was moved from Primitive to Types in version 0.6.3.0.

sizeOfType :: Prim a => Int

The size of values of type a in bytes. This has to be used with TypeApplications: sizeOfType @a.

>>> :set -XTypeApplications
>>> import Data.Int (Int32)
>>> sizeOfType @Int32
4

Since: 0.9.0.0

alignment :: Prim a => a -> Int

The alignment of values of type a in bytes. The argument is not used.

It is recommended to use alignmentOfType instead.

This function has existed since 0.1, but was moved from Primitive to Types in version 0.6.3.0.

alignmentOfType :: Prim a => Int

The alignment of values of type a in bytes. This has to be used with TypeApplications: alignmentOfType @a.

Since: 0.9.0.0

defaultSetByteArray# :: Prim a => MutableByteArray# s -> Int# -> Int# -> a -> State# s -> State# s

An implementation of setByteArray# that calls writeByteArray# to set each element. This is helpful when writing a Prim instance for a multi-word data type for which there is no CPU-accelerated way to broadcast a value to contiguous memory. It is typically used alongside defaultSetOffAddr#. For example:

data Trip = Trip Int Int Int

instance Prim Trip
  sizeOfType# _ = 3# *# sizeOfType# (proxy# :: Proxy# Int)
  alignmentOfType# _ = alignmentOfType# (proxy# :: Proxy# Int)
  indexByteArray# arr# i# = ...
  readByteArray# arr# i# = ...
  writeByteArray# arr# i# (Trip a b c) =
    \s0 -> case writeByteArray# arr# (3# *# i#) a s0 of
       s1 -> case writeByteArray# arr# ((3# *# i#) +# 1#) b s1 of
         s2 -> case writeByteArray# arr# ((3# *# i#) +# 2# ) c s2 of
           s3 -> s3
  setByteArray# = defaultSetByteArray#
  indexOffAddr# addr# i# = ...
  readOffAddr# addr# i# = ...
  writeOffAddr# addr# i# (Trip a b c) =
    \s0 -> case writeOffAddr# addr# (3# *# i#) a s0 of
       s1 -> case writeOffAddr# addr# ((3# *# i#) +# 1#) b s1 of
         s2 -> case writeOffAddr# addr# ((3# *# i#) +# 2# ) c s2 of
           s3 -> s3
  setOffAddr# = defaultSetOffAddr#

defaultSetOffAddr# :: Prim a => Addr# -> Int# -> Int# -> a -> State# s -> State# s

An implementation of setOffAddr# that calls writeOffAddr# to set each element. The documentation of defaultSetByteArray# provides an example of how to use this.

newtype PrimStorable a

Newtype that uses a Prim instance to give rise to a Storable instance. This type is intended to be used with the DerivingVia extension available in GHC 8.6 and up. For example, consider a user-defined Prim instance for a multi-word data type.

data Uuid = Uuid Word64 Word64
  deriving Storable via (PrimStorable Uuid)
instance Prim Uuid where ...

Writing the Prim instance is tedious and unavoidable, but the Storable instance comes for free once the Prim instance is written.

Constructors

PrimStorable

Fields

getPrimStorable :: a

data Ptr a

Constructors

Ptr Addr#