Control.Monad.Primitive

Description

Primitive state-transformer monads

Documentation

class Monad m => PrimMonad m where

Class of monads which can perform primitive state-transformer actions

Minimal complete definition

primitive

Associated Types

type PrimState m

State token type

Methods

primitive :: (State# (PrimState m) -> (#State# (PrimState m), a#)) -> m a

Execute a primitive operation

data RealWorld :: *

RealWorld is deeply magical. It is primitive, but it is not unlifted (hence ptrArg). We never manipulate values of type RealWorld; it's only used in the type system, to parameterise State#.

primitive_ :: PrimMonad m => (State# (PrimState m) -> State# (PrimState m)) -> m ()

Execute a primitive operation with no result

class PrimMonad m => PrimBase m where

Class of primitive monads for state-transformer actions.

Unlike PrimMonad, this typeclass requires that the Monad be fully expressed as a state transformer, therefore disallowing other monad transformers on top of the base IO or ST.

Minimal complete definition

internal

Methods

internal :: m a -> State# (PrimState m) -> (#State# (PrimState m), a#)

Expose the internal structure of the monad

liftPrim :: (PrimBase m1, PrimMonad m2, PrimState m1 ~ PrimState m2) => m1 a -> m2 a

Lifts a PrimBase into another PrimMonad with the same underlying state token type.

primToPrim :: (PrimBase m1, PrimMonad m2, PrimState m1 ~ PrimState m2) => m1 a -> m2 a

Convert a PrimBase to another monad with the same state token.

primToIO :: (PrimBase m, PrimState m ~ RealWorld) => m a -> IO a

Convert a PrimBase with a RealWorld state token to IO

primToST :: PrimBase m => m a -> ST (PrimState m) a

Convert a PrimBase to ST

ioToPrim :: (PrimMonad m, PrimState m ~ RealWorld) => IO a -> m a

Convert an IO action to a PrimMonad.

stToPrim :: PrimMonad m => ST (PrimState m) a -> m a

Convert an ST action to a PrimMonad.

unsafePrimToPrim :: (PrimBase m1, PrimMonad m2) => m1 a -> m2 a

Convert a PrimBase to another monad with a possibly different state token. This operation is highly unsafe!

unsafePrimToIO :: PrimBase m => m a -> IO a

Convert any PrimBase to IO. This operation is highly unsafe!

unsafePrimToST :: PrimBase m => m a -> ST s a

Convert any PrimBase to ST with an arbitrary state token. This operation is highly unsafe!

unsafeIOToPrim :: PrimMonad m => IO a -> m a

Convert an IO action to any PrimMonad. This operation is highly unsafe!

unsafeSTToPrim :: PrimMonad m => ST s a -> m a

Convert an ST action with an arbitraty state token to any PrimMonad. This operation is highly unsafe!

unsafeInlinePrim :: PrimBase m => m a -> a

unsafeInlineIO :: IO a -> a

unsafeInlineST :: ST s a -> a

touch :: PrimMonad m => a -> m ()

evalPrim :: forall a m. PrimMonad m => a -> m a

Create an action to force a value; generalizes evaluate