Data.Primitive.Array

Description

Primitive arrays of boxed values.

Documentation

data Array a

Boxed arrays

Constructors

Array

Fields

array# :: Array# a

data MutableArray s a

Mutable boxed arrays associated with a primitive state token.

Constructors

MutableArray

Fields

marray# :: MutableArray# s a

newArray :: PrimMonad m => Int -> a -> m (MutableArray (PrimState m) a)

Create a new mutable array of the specified size and initialise all elements with the given value.

readArray :: PrimMonad m => MutableArray (PrimState m) a -> Int -> m a

Read a value from the array at the given index.

writeArray :: PrimMonad m => MutableArray (PrimState m) a -> Int -> a -> m ()

Write a value to the array at the given index.

indexArray :: Array a -> Int -> a

Read a value from the immutable array at the given index.

indexArrayM :: Monad m => Array a -> Int -> m a

Monadically read a value from the immutable array at the given index. This allows us to be strict in the array while remaining lazy in the read element which is very useful for collective operations. Suppose we want to copy an array. We could do something like this:

copy marr arr ... = do ...
                       writeArray marr i (indexArray arr i) ...
                       ...

But since primitive arrays are lazy, the calls to indexArray will not be evaluated. Rather, marr will be filled with thunks each of which would retain a reference to arr. This is definitely not what we want!

With indexArrayM, we can instead write

copy marr arr ... = do ...
                       x <- indexArrayM arr i
                       writeArray marr i x
                       ...

Now, indexing is executed immediately although the returned element is still not evaluated.

freezeArray

Arguments

:: PrimMonad m

=> MutableArray (PrimState m) a
source

-> Int
offset

-> Int
length

-> m (Array a)

Create an immutable copy of a slice of an array.

This operation makes a copy of the specified section, so it is safe to continue using the mutable array afterward.

thawArray

Arguments

:: PrimMonad m

=> Array a
source

-> Int
offset

-> Int
length

-> m (MutableArray (PrimState m) a)

Create a mutable array from a slice of an immutable array.

This operation makes a copy of the specified slice, so it is safe to use the immutable array afterward.

unsafeFreezeArray :: PrimMonad m => MutableArray (PrimState m) a -> m (Array a)

Convert a mutable array to an immutable one without copying. The array should not be modified after the conversion.

unsafeThawArray :: PrimMonad m => Array a -> m (MutableArray (PrimState m) a)

Convert an immutable array to an mutable one without copying. The immutable array should not be used after the conversion.

sameMutableArray :: MutableArray s a -> MutableArray s a -> Bool

Check whether the two arrays refer to the same memory block.

copyArray

Arguments

:: PrimMonad m

=> MutableArray (PrimState m) a
destination array

-> Int
offset into destination array

-> Array a
source array

-> Int
offset into source array

-> Int
number of elements to copy

-> m ()

Copy a slice of an immutable array to a mutable array.

copyMutableArray

Arguments

:: PrimMonad m

=> MutableArray (PrimState m) a
destination array

-> Int
offset into destination array

-> MutableArray (PrimState m) a
source array

-> Int
offset into source array

-> Int
number of elements to copy

-> m ()

Copy a slice of a mutable array to another array. The two arrays may not be the same.

cloneArray

Arguments

:: Array a
source array

-> Int
offset into destination array

-> Int
number of elements to copy

-> Array a

Return a newly allocated Array with the specified subrange of the provided Array. The provided Array should contain the full subrange specified by the two Ints, but this is not checked.

cloneMutableArray

Arguments

:: PrimMonad m

=> MutableArray (PrimState m) a
source array

-> Int
offset into destination array

-> Int
number of elements to copy

-> m (MutableArray (PrimState m) a)

Return a newly allocated MutableArray. with the specified subrange of the provided MutableArray. The provided MutableArray should contain the full subrange specified by the two Ints, but this is not checked.

sizeofArray :: Array a -> Int

sizeofMutableArray :: MutableArray s a -> Int

fromListN :: IsList l => Int -> [Item l] -> l

The fromListN function takes the input list's length as a hint. Its behaviour should be equivalent to fromList. The hint can be used to construct the structure l more efficiently compared to fromList. If the given hint does not equal to the input list's length the behaviour of fromListN is not specified.

fromList :: IsList l => [Item l] -> l

The fromList function constructs the structure l from the given list of Item l