Data.Primitive.ByteArray

Description

Primitive operations on ByteArrays

Types

data ByteArray

Byte arrays

Constructors

ByteArray ByteArray#

data MutableByteArray s

Mutable byte arrays associated with a primitive state token

Constructors

MutableByteArray (MutableByteArray# s)

data ByteArray# :: TYPE UnliftedRep

data MutableByteArray# (a :: *) :: * -> TYPE UnliftedRep

Allocation

newByteArray :: PrimMonad m => Int -> m (MutableByteArray (PrimState m))

Create a new mutable byte array of the specified size in bytes.

newPinnedByteArray :: PrimMonad m => Int -> m (MutableByteArray (PrimState m))

Create a pinned byte array of the specified size in bytes. The garbage collector is guaranteed not to move it.

newAlignedPinnedByteArray :: PrimMonad m => Int -> Int -> m (MutableByteArray (PrimState m))

Create a pinned byte array of the specified size in bytes and with the give alignment. The garbage collector is guaranteed not to move it.

Element access

readByteArray :: (Prim a, PrimMonad m) => MutableByteArray (PrimState m) -> Int -> m a

Read a primitive value from the byte array. The offset is given in elements of type a rather than in bytes.

writeByteArray :: (Prim a, PrimMonad m) => MutableByteArray (PrimState m) -> Int -> a -> m ()

Write a primitive value to the byte array. The offset is given in elements of type a rather than in bytes.

indexByteArray :: Prim a => ByteArray -> Int -> a

Read a primitive value from the byte array. The offset is given in elements of type a rather than in bytes.

Folding

foldrByteArray :: forall a b. Prim a => (a -> b -> b) -> b -> ByteArray -> b

Right-fold over the elements of a ByteArray.

Freezing and thawing

unsafeFreezeByteArray :: PrimMonad m => MutableByteArray (PrimState m) -> m ByteArray

Convert a mutable byte array to an immutable one without copying. The array should not be modified after the conversion.

unsafeThawByteArray :: PrimMonad m => ByteArray -> m (MutableByteArray (PrimState m))

Convert an immutable byte array to a mutable one without copying. The original array should not be used after the conversion.

Block operations

copyByteArray

Arguments

:: PrimMonad m

=> MutableByteArray (PrimState m)
destination array

-> Int
offset into destination array

-> ByteArray
source array

-> Int
offset into source array

-> Int
number of bytes to copy

-> m ()

Copy a slice of an immutable byte array to a mutable byte array.

copyMutableByteArray

Arguments

:: PrimMonad m

=> MutableByteArray (PrimState m)
destination array

-> Int
offset into destination array

-> MutableByteArray (PrimState m)
source array

-> Int
offset into source array

-> Int
number of bytes to copy

-> m ()

Copy a slice of a mutable byte array into another array. The two slices may not overlap.

moveByteArray

Arguments

:: PrimMonad m

=> MutableByteArray (PrimState m)
destination array

-> Int
offset into destination array

-> MutableByteArray (PrimState m)
source array

-> Int
offset into source array

-> Int
number of bytes to copy

-> m ()

Copy a slice of a mutable byte array into another, potentially overlapping array.

setByteArray

Arguments

:: (Prim a, PrimMonad m)

=> MutableByteArray (PrimState m)
array to fill

-> Int
offset into array

-> Int
number of values to fill

-> a
value to fill with

-> m ()

Fill a slice of a mutable byte array with a value. The offset and length are given in elements of type a rather than in bytes.

fillByteArray

Arguments

:: PrimMonad m

=> MutableByteArray (PrimState m)
array to fill

-> Int
offset into array

-> Int
number of bytes to fill

-> Word8
byte to fill with

-> m ()

Fill a slice of a mutable byte array with a byte.

Information

sizeofByteArray :: ByteArray -> Int

Size of the byte array in bytes.

sizeofMutableByteArray :: MutableByteArray s -> Int

Size of the mutable byte array in bytes.

sameMutableByteArray :: MutableByteArray s -> MutableByteArray s -> Bool

Check if the two arrays refer to the same memory block.

byteArrayContents :: ByteArray -> Addr

Yield a pointer to the array's data. This operation is only safe on pinned byte arrays allocated by newPinnedByteArray or newAlignedPinnedByteArray.

mutableByteArrayContents :: MutableByteArray s -> Addr

Yield a pointer to the array's data. This operation is only safe on pinned byte arrays allocated by newPinnedByteArray or newAlignedPinnedByteArray.