Data.Primitive.ByteArray

Description

Primitive operations on byte arrays. Most functions in this module include an element type in their type signature and interpret the unit for offsets and lengths as that element. A few functions (e.g. copyByteArray, freezeByteArray) do not include an element type. Such functions interpret offsets and lengths as units of 8-bit words.

Types

data ByteArray

Byte arrays.

Constructors

ByteArray ByteArray#

data MutableByteArray s

Mutable byte arrays associated with a primitive state token.

Constructors

MutableByteArray (MutableByteArray# s)

data ByteArray# :: UnliftedType

A boxed, unlifted datatype representing a region of raw memory in the garbage-collected heap, which is not scanned for pointers during garbage collection.

It is created by freezing a 'MutableByteArray#' with 'unsafeFreezeByteArray#'. Freezing is essentially a no-op, as MutableByteArray# and ByteArray# share the same heap structure under the hood.

The immutable and mutable variants are commonly used for scenarios requiring high-performance data structures, like Text, Primitive Vector, Unboxed Array, and ShortByteString.

Another application of fundamental importance is 'Integer', which is backed by 'ByteArray#'.

The representation on the heap of a Byte Array is:

+------------+-----------------+-----------------------+
|            |                 |                       |
|   HEADER   | SIZE (in bytes) |       PAYLOAD         |
|            |                 |                       |
+------------+-----------------+-----------------------+

To obtain a pointer to actual payload (e.g., for FFI purposes) use 'byteArrayContents#' or 'mutableByteArrayContents#'.

Alternatively, enabling the UnliftedFFITypes extension allows to mention 'ByteArray#' and 'MutableByteArray#' in FFI type signatures directly.

data MutableByteArray# a :: UnliftedType

A mutable ByteAray#. It can be created in three ways:

'newByteArray#': Create an unpinned array.

'newPinnedByteArray#': This will create a pinned array,

'newAlignedPinnedByteArray#': This will create a pinned array, with a custom alignment.

Unpinned arrays can be moved around during garbage collection, so you must not store or pass pointers to these values if there is a chance for the garbage collector to kick in. That said, even unpinned arrays can be passed to unsafe FFI calls, because no garbage collection happens during these unsafe calls (see Guaranteed Call Safety in the GHC Manual). For safe FFI calls, byte arrays must be not only pinned, but also kept alive by means of the keepAlive# function for the duration of a call (that's because garbage collection cannot move a pinned array, but is free to scrap it altogether).

Allocation

newByteArray :: PrimMonad m => Int -> m (MutableByteArray (PrimState m))

Create a new mutable byte array of the specified size in bytes.

Note: this function does not check if the input is non-negative.

newPinnedByteArray :: PrimMonad m => Int -> m (MutableByteArray (PrimState m))

Create a pinned byte array of the specified size in bytes. The garbage collector is guaranteed not to move it.

Note: this function does not check if the input is non-negative.

newAlignedPinnedByteArray

Arguments

:: PrimMonad m

=> Int
size

-> Int
alignment

-> m (MutableByteArray (PrimState m))

Create a pinned byte array of the specified size in bytes and with the given alignment. The garbage collector is guaranteed not to move it.

Note: this function does not check if the input is non-negative.

resizeMutableByteArray :: PrimMonad m => MutableByteArray (PrimState m) -> Int -> m (MutableByteArray (PrimState m))

Resize a mutable byte array. The new size is given in bytes.

This will either resize the array in-place or, if not possible, allocate the contents into a new, unpinned array and copy the original array's contents.

To avoid undefined behaviour, the original MutableByteArray shall not be accessed anymore after a resizeMutableByteArray has been performed. Moreover, no reference to the old one should be kept in order to allow garbage collection of the original MutableByteArray in case a new MutableByteArray had to be allocated.

Since: 0.6.4.0

shrinkMutableByteArray

Arguments

:: PrimMonad m

=> MutableByteArray (PrimState m)

-> Int
new size

-> m ()

Shrink a mutable byte array. The new size is given in bytes. It must be smaller than the old size. The array will be resized in place.

Since: 0.7.1.0

Element access

readByteArray :: (Prim a, PrimMonad m) => MutableByteArray (PrimState m) -> Int -> m a

Read a primitive value from the byte array. The offset is given in elements of type a rather than in bytes.

Note: this function does not do bounds checking.

writeByteArray :: (Prim a, PrimMonad m) => MutableByteArray (PrimState m) -> Int -> a -> m ()

Write a primitive value to the byte array. The offset is given in elements of type a rather than in bytes.

Note: this function does not do bounds checking.

indexByteArray :: Prim a => ByteArray -> Int -> a

Read a primitive value from the byte array. The offset is given in elements of type a rather than in bytes.

Note: this function does not do bounds checking.

Constructing

emptyByteArray :: ByteArray

The empty ByteArray.

byteArrayFromList :: Prim a => [a] -> ByteArray

Create a ByteArray from a list.

byteArrayFromList xs = byteArrayFromListN (length xs) xs

byteArrayFromListN :: Prim a => Int -> [a] -> ByteArray

Create a ByteArray from a list of a known length. If the length of the list does not match the given length, this throws an exception.

Folding

foldrByteArray :: forall a b. Prim a => (a -> b -> b) -> b -> ByteArray -> b

Right-fold over the elements of a ByteArray.

Comparing

compareByteArrays

Arguments

:: ByteArray
array A

-> Int
offset A, given in bytes

-> ByteArray
array B

-> Int
offset B, given in bytes

-> Int
length of the slice, given in bytes

-> Ordering

Lexicographic comparison of equal-length slices into two byte arrays. This wraps the compareByteArrays# primop, which wraps memcmp.

Freezing and thawing

freezeByteArray

Arguments

:: PrimMonad m

=> MutableByteArray (PrimState m)
source

-> Int
offset in bytes

-> Int
length in bytes

-> m ByteArray

Create an immutable copy of a slice of a byte array. The offset and length are given in bytes.

This operation makes a copy of the specified section, so it is safe to continue using the mutable array afterward.

Note: The provided array should contain the full subrange specified by the two Ints, but this is not checked.

thawByteArray

Arguments

:: PrimMonad m

=> ByteArray
source

-> Int
offset in bytes

-> Int
length in bytes

-> m (MutableByteArray (PrimState m))

Create a mutable byte array from a slice of an immutable byte array. The offset and length are given in bytes.

This operation makes a copy of the specified slice, so it is safe to use the immutable array afterward.

Note: The provided array should contain the full subrange specified by the two Ints, but this is not checked.

Since: 0.7.2.0

runByteArray :: (forall s. ST s (MutableByteArray s)) -> ByteArray

Execute the monadic action and freeze the resulting array.

runByteArray m = runST $ m >>= unsafeFreezeByteArray

unsafeFreezeByteArray :: PrimMonad m => MutableByteArray (PrimState m) -> m ByteArray

Convert a mutable byte array to an immutable one without copying. The array should not be modified after the conversion.

unsafeThawByteArray :: PrimMonad m => ByteArray -> m (MutableByteArray (PrimState m))

Convert an immutable byte array to a mutable one without copying. The original array should not be used after the conversion.

Block operations

copyByteArray

Arguments

:: PrimMonad m

=> MutableByteArray (PrimState m)
destination array

-> Int
offset into destination array

-> ByteArray
source array

-> Int
offset into source array

-> Int
number of bytes to copy

-> m ()

Copy a slice of an immutable byte array to a mutable byte array.

Note: this function does not do bounds or overlap checking.

copyMutableByteArray

Arguments

:: PrimMonad m

=> MutableByteArray (PrimState m)
destination array

-> Int
offset into destination array

-> MutableByteArray (PrimState m)
source array

-> Int
offset into source array

-> Int
number of bytes to copy

-> m ()

Copy a slice of a mutable byte array into another array. The two slices may not overlap.

Note: this function does not do bounds or overlap checking.

copyByteArrayToPtr

Arguments

:: forall m a. (PrimMonad m, Prim a)

=> Ptr a
destination

-> ByteArray
source array

-> Int
offset into source array, interpreted as elements of type a

-> Int
number of elements to copy

-> m ()

Copy a slice of a byte array to an unmanaged pointer address. These must not overlap. The offset and length are given in elements, not in bytes.

Note: this function does not do bounds or overlap checking.

Since: 0.7.1.0

copyMutableByteArrayToPtr

Arguments

:: forall m a. (PrimMonad m, Prim a)

=> Ptr a
destination

-> MutableByteArray (PrimState m)
source array

-> Int
offset into source array, interpreted as elements of type a

-> Int
number of elements to copy

-> m ()

Copy a slice of a mutable byte array to an unmanaged pointer address. These must not overlap. The offset and length are given in elements, not in bytes.

Note: this function does not do bounds or overlap checking.

Since: 0.7.1.0

copyByteArrayToAddr

Arguments

:: PrimMonad m

=> Ptr Word8
destination

-> ByteArray
source array

-> Int
offset into source array

-> Int
number of bytes to copy

-> m ()

Copy a slice of a byte array to an unmanaged address. These must not overlap.

Note: This function is just copyByteArrayToPtr where a is Word8.

Since: 0.6.4.0

copyMutableByteArrayToAddr

Arguments

:: PrimMonad m

=> Ptr Word8
destination

-> MutableByteArray (PrimState m)
source array

-> Int
offset into source array

-> Int
number of bytes to copy

-> m ()

Copy a slice of a mutable byte array to an unmanaged address. These must not overlap.

Note: This function is just copyMutableByteArrayToPtr where a is Word8.

Since: 0.6.4.0

moveByteArray

Arguments

:: PrimMonad m

=> MutableByteArray (PrimState m)
destination array

-> Int
offset into destination array

-> MutableByteArray (PrimState m)
source array

-> Int
offset into source array

-> Int
number of bytes to copy

-> m ()

Copy a slice of a mutable byte array into another, potentially overlapping array.

setByteArray

Arguments

:: (Prim a, PrimMonad m)

=> MutableByteArray (PrimState m)
array to fill

-> Int
offset into array

-> Int
number of values to fill

-> a
value to fill with

-> m ()

Fill a slice of a mutable byte array with a value. The offset and length are given in elements of type a rather than in bytes.

Note: this function does not do bounds checking.

fillByteArray

Arguments

:: PrimMonad m

=> MutableByteArray (PrimState m)
array to fill

-> Int
offset into array

-> Int
number of bytes to fill

-> Word8
byte to fill with

-> m ()

Fill a slice of a mutable byte array with a byte.

Note: this function does not do bounds checking.

cloneByteArray

Arguments

:: ByteArray
source array

-> Int
offset into destination array

-> Int
number of bytes to copy

-> ByteArray

Return a newly allocated array with the specified subrange of the provided array. The provided array should contain the full subrange specified by the two Ints, but this is not checked.

cloneMutableByteArray

Arguments

:: PrimMonad m

=> MutableByteArray (PrimState m)
source array

-> Int
offset into destination array

-> Int
number of bytes to copy

-> m (MutableByteArray (PrimState m))

Return a newly allocated mutable array with the specified subrange of the provided mutable array. The provided mutable array should contain the full subrange specified by the two Ints, but this is not checked.

Information

sizeofByteArray :: ByteArray -> Int

Size of the byte array in bytes.

sizeofMutableByteArray :: MutableByteArray s -> Int

Size of the mutable byte array in bytes. This function's behavior is undefined if resizeMutableByteArray is ever called on the mutable byte array given as the argument. Consequently, use of this function is discouraged. Prefer getSizeofMutableByteArray, which ensures correct sequencing in the presence of resizing.

getSizeofMutableByteArray :: PrimMonad m => MutableByteArray (PrimState m) -> m Int

Get the size of a byte array in bytes. Unlike sizeofMutableByteArray, this function ensures sequencing in the presence of resizing.

sameMutableByteArray :: MutableByteArray s -> MutableByteArray s -> Bool

Check if the two arrays refer to the same memory block.

isByteArrayPinned :: ByteArray -> Bool

Check whether or not the byte array is pinned. Pinned byte arrays cannot be moved by the garbage collector. It is safe to use byteArrayContents on such byte arrays. This function is only available when compiling with GHC 8.2 or newer.

Since: 0.6.4.0

isMutableByteArrayPinned :: MutableByteArray s -> Bool

Check whether or not the mutable byte array is pinned. This function is only available when compiling with GHC 8.2 or newer.

Since: 0.6.4.0

byteArrayContents :: ByteArray -> Ptr Word8

Yield a pointer to the array's data. This operation is only safe on pinned byte arrays allocated by newPinnedByteArray or newAlignedPinnedByteArray.

mutableByteArrayContents :: MutableByteArray s -> Ptr Word8

Yield a pointer to the array's data. This operation is only safe on pinned byte arrays allocated by newPinnedByteArray or newAlignedPinnedByteArray.