Data.Primitive.SmallArray

Description

Small arrays are boxed (im)mutable arrays.

The underlying structure of the Array type contains a card table, allowing segments of the array to be marked as having been mutated. This allows the garbage collector to only re-traverse segments of the array that have been marked during certain phases, rather than having to traverse the entire array.

SmallArray lacks this table. This means that it takes up less memory and has slightly faster writes. It is also more efficient during garbage collection so long as the card table would have a single entry covering the entire array. These advantages make them suitable for use as arrays that are known to be small.

The card size is 128, so for uses much larger than that, Array would likely be superior.

The underlying type, SmallArray#, was introduced in GHC 7.10, so prior to that version, this module simply implements small arrays as Array.

Documentation

data SmallArray a

Constructors

SmallArray (SmallArray# a)

data SmallMutableArray s a

Constructors

SmallMutableArray (SmallMutableArray# s a)

newSmallArray

Arguments

:: PrimMonad m

=> Int
size

-> a
initial contents

-> m (SmallMutableArray (PrimState m) a)

Create a new small mutable array.

readSmallArray

Arguments

:: PrimMonad m

=> SmallMutableArray (PrimState m) a
array

-> Int
index

-> m a

Read the element at a given index in a mutable array.

writeSmallArray

Arguments

:: PrimMonad m

=> SmallMutableArray (PrimState m) a
array

-> Int
index

-> a
new element

-> m ()

Write an element at the given idex in a mutable array.

copySmallArray

Arguments

:: PrimMonad m

=> SmallMutableArray (PrimState m) a
destination

-> Int
destination offset

-> SmallArray a
source

-> Int
source offset

-> Int
length

-> m ()

Copy a slice of an immutable array into a mutable array.

copySmallMutableArray

Arguments

:: PrimMonad m

=> SmallMutableArray (PrimState m) a
destination

-> Int
destination offset

-> SmallMutableArray (PrimState m) a
source

-> Int
source offset

-> Int
length

-> m ()

Copy a slice of one mutable array into another.

indexSmallArray

Arguments

:: SmallArray a
array

-> Int
index

-> a

Look up an element in an immutable array.

indexSmallArrayM

Arguments

:: Monad m

=> SmallArray a
array

-> Int
index

-> m a

Look up an element in an immutable array.

The purpose of returning a result using a monad is to allow the caller to avoid retaining references to the array. Evaluating the return value will cause the array lookup to be performed, even though it may not require the element of the array to be evaluated (which could throw an exception). For instance:

data Box a = Box a
...

f sa = case indexSmallArrayM sa 0 of
  Box x -> ...

x is not a closure that references sa as it would be if we instead wrote:

let x = indexSmallArray sa 0

And does not prevent sa from being garbage collected.

Note that Identity is not adequate for this use, as it is a newtype, and cannot be evaluated without evaluating the element.

indexSmallArray## :: SmallArray a -> Int -> (#a#)

Read a value from the immutable array at the given index, returning the result in an unboxed unary tuple. This is currently used to implement folds.

cloneSmallArray

Arguments

:: SmallArray a
source

-> Int
offset

-> Int
length

-> SmallArray a

Create a copy of a slice of an immutable array.

cloneSmallMutableArray

Arguments

:: PrimMonad m

=> SmallMutableArray (PrimState m) a
source

-> Int
offset

-> Int
length

-> m (SmallMutableArray (PrimState m) a)

Create a copy of a slice of a mutable array.

freezeSmallArray

Arguments

:: PrimMonad m

=> SmallMutableArray (PrimState m) a
source

-> Int
offset

-> Int
length

-> m (SmallArray a)

Create an immutable array corresponding to a slice of a mutable array.

This operation copies the portion of the array to be frozen.

unsafeFreezeSmallArray :: PrimMonad m => SmallMutableArray (PrimState m) a -> m (SmallArray a)

Render a mutable array immutable.

This operation performs no copying, so care must be taken not to modify the input array after freezing.

thawSmallArray

Arguments

:: PrimMonad m

=> SmallArray a
source

-> Int
offset

-> Int
length

-> m (SmallMutableArray (PrimState m) a)

Create a mutable array corresponding to a slice of an immutable array.

This operation copies the portion of the array to be thawed.

runSmallArray :: (forall s. ST s (SmallMutableArray s a)) -> SmallArray a

unsafeThawSmallArray :: PrimMonad m => SmallArray a -> m (SmallMutableArray (PrimState m) a)

Render an immutable array mutable.

This operation performs no copying, so care must be taken with its use.

sizeofSmallArray :: SmallArray a -> Int

sizeofSmallMutableArray :: SmallMutableArray s a -> Int

smallArrayFromList :: [a] -> SmallArray a

Create a SmallArray from a list.

smallArrayFromListN :: Int -> [a] -> SmallArray a

Create a SmallArray from a list of a known length. If the length of the list does not match the given length, this throws an exception.

mapSmallArray' :: (a -> b) -> SmallArray a -> SmallArray b

Strict map over the elements of the array.

traverseSmallArrayP :: PrimMonad m => (a -> m b) -> SmallArray a -> m (SmallArray b)

This is the fastest, most straightforward way to traverse an array, but it only works correctly with a sufficiently "affine" PrimMonad instance. In particular, it must only produce *one* result array. ListT-transformed monads, for example, will not work right at all.